template: remove magic behaviour of set('debug', ...)
Technically this is documented, but we're not using it anywhere.  If you
uses any of these names:

    debug|active_theme|default_theme|current_view|layout|parse_views|data|
    blocks|message|theme_paths|site_path|orig_view_path|ci

they will set class properties of Template, instead of adding a variable to
the view data.  The name "data" in particular seems ready to cause many
man-hours of debugging fun.

There's no alternative setter for $debug, so I just made it public.
Remember that without this change, *all* the properties of the Template
class are effectively public.