emailer preview: view as html, instead of source code
The emails are sent as HTML, but if you use the "preview" (e.g. on an
activation email after forcing $queue_emails = TRUE), it was forcing
you to read the source code.

This could also be misleading.  Say you have a message that _looks_ like
plain text, but is actually HTML.  It could be invalid HTML due to a naked
'&' character, but you wouldn't realize that.